Add KVCache transpose for long sequence && tuned comm for faster Addreduce

+++ src/utils/matmul_helper.h
@@ -1067,7 +1067,7 @@ public:
             if constexpr (std::is_same_v<InT, bfloat16_t>) {
                 TimeLine t("onednn_amx_sgemm_f32bf16f32_compute_residential");
 #pragma omp parallel for collapse(2)
-                for (int i = 0; i < M; ++i) {
+                for (uint64_t i = 0; i < M; ++i) {
                     for (int j = 0; j < N; ++j) {
                         auto remain = N - j;
                         __mmask16 mask = (remain >= 16 ? 0xffff : (1 << remain) - 1);
@@ -1082,7 +1082,7 @@ public:
                 if (M > AMXThresholdM) {
                     TimeLine t("onednn_amx_sgemm_f32bf16f32_compute_residential");
 #pragma omp parallel for collapse(2)
-                    for (int i = 0; i < M; ++i) {
+                    for (uint64_t i = 0; i < M; ++i) {
                         for (int j = 0; j < N; ++j) {
                             res[i * ldres + j] = res[i * ldres + j] * gamma;
                         }
@@ -1624,7 +1624,7 @@ private:
         if (C == res) {
             scale_mem = memory(scale_md, *engine);
 #pragma omp parallel for
-            for (int i = 0; i < M; ++i) {
+            for (uint64_t i = 0; i < M; ++i) {
                 memcpy((Tin *)scale_mem.get_data_handle() + i * N, res + i * ldres, N * sizeof(Tin));
             }
         } else {
